Moved recommendation algorithm from user datum MVC to daily questionnaire MVC
