+++ app/views/home/_header.html.erb
@@ -1,7 +1,7 @@
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
   <div class="container-fluid">
     <%= link_to 'Health App', root_path, class:"navbar-brand" %>
-    <%= link_to 'Recommendations', '/recommendations' %>
+    <%= link_to 'Daily Recommendations', '/daily_recommendations' %>
     <%# <%= link_to "Today's Questionnaire", "/new_daily_questionnaire" %>
     <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
       <span class="navbar-toggler-icon"></span>
